+++ Cards/Status/AscendersBane.cs
@@ -9,6 +9,6 @@ namespace StS
         public override TargetType TargetType => TargetType.Player;
         public override int CiCanCallEnergyCost(int upgradeCount) => int.MaxValue;
         internal override bool Ethereal(int upgradeCount) => true;
-        internal override void Play(EffectSet ef, Player player, IEnemy enemy, int upgradeCount, IList<CardInstance> targets = null, Deck deck = null, int? key = null) { }
+        internal override void Play(EffectSet ef, Player player, IEnemy enemy, int upgradeCount, IList<CardInstance> targets = null, Deck deck = null, long? key = null) { }
     }
 }
